+++ _universum/modules/vcs/perforce_vcs.py
@@ -1,5 +1,6 @@
 # -*- coding: UTF-8 -*-
 
+import json
 import os
 import shutil
 import warnings
@@ -7,6 +8,7 @@ import warnings
 import sh
 
 from ...modules.artifact_collector import ArtifactCollector
+from ...modules.api_support import ApiSupport
 from ...modules.reporter import Reporter
 from ...lib.ci_exception import CriticalCiException, SilentAbortException
 from ...lib.gravity import Dependency
@@ -223,6 +225,7 @@ class PerforceWithMappings(PerforceVcs):
 class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
     swarm_factory = Dependency(Swarm)
     artifacts_factory = Dependency(ArtifactCollector)
+    api_support_factory = Dependency(ApiSupport)
     reporter_factory = Dependency(Reporter)
 
     @staticmethod
@@ -257,6 +260,7 @@ class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
         super(PerforceMainVcs, self).__init__(*args, **kwargs)
 
         self.artifacts = self.artifacts_factory()
+        self.api_support = self.api_support_factory()
         self.reporter = self.reporter_factory()
         # self.swarm is initialized by self.code_review()
         self.swarm = None
@@ -277,38 +281,14 @@ class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
         self.swarm = self.swarm_factory(self.settings.user, self.settings.password)
         return self.swarm
 
-    def parse_description(self, cl_number):
+    def get_related_cls(self, cl_number):
+        cl_list = [cl_number]
         description = self.p4.run_describe(cl_number)[0]
         for entry in description['desc'].splitlines():
             if entry.startswith("[Related change IDs]"):
-                cl_list = [number.strip() for number in entry.strip("[Related change IDs]").split(",")]
-                if cl_number in cl_list:
-                    return cl_list
-                self.reporter.add_block_to_report(self.structure.get_current_block())
-                self.structure.fail_current_block("Current CL is not in related list!")
+                cl_list.extend([number.strip() for number in entry.strip("[Related change IDs]").split(",")])
 
-        return [cl_number]
-
-    @make_block("Checking that current and master CLs related change IDs are the same", False)
-    def get_related_cls(self, cl_number):
-        cl_list = self.parse_description(cl_number)
-        if not cl_list:
-            return cl_list
-
-        master_cl = cl_list[-1]
-        if master_cl == cl_number:
-            return cl_list
-
-        master_list = self.parse_description(master_cl)
-        if cl_list != master_list:
-            self.reporter.add_block_to_report(self.structure.get_current_block())
-            self.structure.fail_current_block("Related CLs list doesn't match master CL related list!")
-            return cl_list
-
-        self.out.log("Not a master CL, no check needed")
-        self.out.report_build_status("Not a master CL")
-        self.swarm = None
-        raise SilentAbortException(application_exit_code=0)
+        return cl_list
 
     def expand_workspace_parameters(self):
         # Create a list of depots for sync
@@ -335,10 +315,10 @@ class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
                         self.depots.pop(index)
                 self.depots.append({"path": splat_entry[0], "cl": splat_entry[1]})
 
-        # Retrieve list of shelved CLs from "classic" environment variables
+        # Retrieve list of shelved CLs from Swarm and "classic" environment variables
         cls = []
         if self.swarm:
-            swarm_cls = self.get_related_cls(self.swarm.settings.review_id)
+            swarm_cls = self.get_related_cls(self.swarm.settings.change)
             cls.extend(swarm_cls)
         for x in range(1, 6):
             cls.append(os.getenv("SHELVE_CHANGELIST_" + unicode(x)))
@@ -444,9 +424,10 @@ class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
     def check_diff_for_depot(self, depot):
         try:
             p4cmd = sh.Command("p4")
-            result = utils.trim_and_convert_to_unicode(p4cmd("-c", self.settings.client, "-u", self.settings.user,
-                                                             "-P", self.settings.password, "-p", self.settings.port,
-                                                             "diff", depot))
+            diff_result = p4cmd("-c", self.settings.client, "-u", self.settings.user,
+                                "-P", self.settings.password, "-p", self.settings.port,
+                                "diff", depot)
+            result = utils.trim_and_convert_to_unicode(diff_result.stdout)
         except sh.ErrorReturnCode as e:
             for line in e.stderr.splitlines():
                 if not (line.startswith("Librarian checkout")
@@ -457,8 +438,24 @@ class PerforceMainVcs(PerforceWithMappings, base_vcs.BaseDownloadVcs):
             result = utils.trim_and_convert_to_unicode(e.stdout)
         return result
 
+    def calculate_file_diff(self):
+        action_list = {}
+        for entry in self.p4.run_opened():
+            action_list[entry["depotFile"]] = entry["action"]
+        if not action_list:
+            return json.dumps({})
+
+        result = []
+        # Both 'p4 opened' and 'p4 where' entries have same key 'depotFile'
+        for entry in self.p4.run_where(action_list.keys()):
+            result.append({"action": action_list[entry["depotFile"]],
+                           "repo_path": entry["depotFile"],
+                           "local_path": entry["path"]})
+        return json.dumps(result, indent=4)
+
     @make_block("Checking diff")
     def diff(self):
+        self.api_support.add_p4_file_diff(self.calculate_file_diff())
         rep_diff = []
         for depot in self.depots:
             line = depot["path"] + '@' + depot["cl"]